+++ data/resources/CHANGELOG.md
@@ -6,6 +6,7 @@ April 13, 2022:
 - Fix Spell.Stacks to now only consider a duration if a nonzero value is provdied,
   otherwise the stacks check will ignore duration. Also applies to StacksPet.
 - Fix crash in MapObject.cpp (#455).
+- Fix Macro TLO to allow IsTLO and IsVariable while a macro is not running (#452).
 - Add double click to select file in imgui file dialog (#453).
 - Add Group.LowMana to report group member with lowest mana below threshold. Works like
   Group.Injured (#477).